requirements.txt
Flask==3.0.0
yfinance==0.2.40
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
pandas==2.1.4
lxml==5.1.0